Improve UX and styles (see desc)
Added a margin on larger screens, this helps with user retention and
improves readability.

Removed "branded" colors, they have a lower contrast, they are instead
replaced with the soft off-white (from the Catppuccin theme) and black text.

Replaced the font with the default system sans-serif font (allowing for
the user to configure it themselves) this should help with accessibility
(ie. users of OpenDyslexia and Atkinson Hyperlegible).